# Optimize bounds checking's ternary operator use