By supplying a value on both sides of the check, browsers are able to
always assume a value will exist (even if it is bypassed by the thrown
error), and optimize accordingly.