@jbowes 2017-02-11
This is particularly useful for slice manipulation heavy crypto code.

Test code:

```go
package main

import "golang.org/x/crypto/scrypt"

func main() {
    scrypt.Key([]byte("password"), []byte("salt"), 32768, 8, 1, 32)
}
```

Before: 0m10.289s
After: 0m1.387s