@shurcooL 2017-02-11
I tested the go1 benchmarks (#573) before and after this change, and I am seeing quite significant speedups across the board. This was with node version `v7.5.0`. A couple minor slowdowns, but I think they might be within testing variance.

```
benchmark                          old ns/op       new ns/op       delta
BenchmarkBinaryTree17              1974000000      1421000000      -28.01%
BenchmarkFannkuch11                18673000000     19407000000     +3.93%
BenchmarkFmtFprintfEmpty           5470            3823            -30.11%
BenchmarkFmtFprintfString          14200           11180           -21.27%
BenchmarkFmtFprintfInt             13150           9770            -25.70%
BenchmarkFmtFprintfIntInt          21900           17250           -21.23%
BenchmarkFmtFprintfPrefixedInt     16470           13740           -16.58%
BenchmarkFmtFprintfFloat           43000           33020           -23.21%
BenchmarkFmtManyArgs               84700           68200           -19.48%
BenchmarkGobDecode                 728500000       680500000       -6.59%
BenchmarkGobEncode                 777500000       751500000       -3.34%
BenchmarkGzip                      16180000000     13672000000     -15.50%
BenchmarkGunzip                    4273000000      4286000000      +0.30%
BenchmarkJSONEncode                1960000000      1939000000      -1.07%
BenchmarkJSONDecode                3015000000      3103000000      +2.92%
BenchmarkMandelbrot200             3950000         3922000         -0.71%
BenchmarkGoParse                   109000000       109900000       +0.83%
BenchmarkRegexpMatchEasy0_32       5150            5245            +1.84%
BenchmarkRegexpMatchEasy0_1K       19350           19400           +0.26%
BenchmarkRegexpMatchEasy1_32       4570            4610            +0.88%
BenchmarkRegexpMatchEasy1_1K       30760           31980           +3.97%
BenchmarkRegexpMatchMedium_32      9510            9275            -2.47%
BenchmarkRegexpMatchMedium_1K      2270000         2090000         -7.93%
BenchmarkRegexpMatchHard_32        146500          104500          -28.67%
BenchmarkRegexpMatchHard_1K        4260000         2838000         -33.38%
BenchmarkRevcomp                   14344000000     13694000000     -4.53%
BenchmarkTemplate                  4285000000      4224000000      -1.42%
BenchmarkTimeParse                 15000           14980           -0.13%
BenchmarkTimeFormat                49433           48133           -2.63%

benchmark                         old MB/s     new MB/s     speedup
BenchmarkGobDecode                1.05         1.13         1.08x
BenchmarkGobEncode                0.99         1.02         1.03x
BenchmarkGzip                     1.20         1.42         1.18x
BenchmarkGunzip                   4.54         4.53         1.00x
BenchmarkJSONEncode               0.99         1.00         1.01x
BenchmarkJSONDecode               0.64         0.63         0.98x
BenchmarkGoParse                  0.53         0.53         1.00x
BenchmarkRegexpMatchEasy0_32      6.21         6.10         0.98x
BenchmarkRegexpMatchEasy0_1K      52.92        52.78        1.00x
BenchmarkRegexpMatchEasy1_32      7.00         6.94         0.99x
BenchmarkRegexpMatchEasy1_1K      33.29        32.02        0.96x
BenchmarkRegexpMatchMedium_32     0.11         0.11         1.00x
BenchmarkRegexpMatchMedium_1K     0.45         0.49         1.09x
BenchmarkRegexpMatchHard_32       0.22         0.31         1.41x
BenchmarkRegexpMatchHard_1K       0.24         0.36         1.50x
BenchmarkRevcomp                  17.72        18.56        1.05x
BenchmarkTemplate                 0.45         0.46         1.02x
```

/cc @neelance