@jbowes 2017-02-12
> This is a really clever and useful change. Thanks for contributing!

Thanks! It was my pleasure :)